pyproject.toml
[project]
name = "fediaf-verifier"
version = "1.0.0"
description = "Pet food label verifier against FEDIAF 2021 + EU 767/2009"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.85.0",
    "streamlit>=1.55.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "Pillow>=12.1.1",
    "loguru>=0.7.0",
]

[project.optional-dependencies]
docx-convert = [
    "pdf2image>=1.17.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.0",
    "pyright>=1.1.0",
    "pytest-mock>=3.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fediaf_verifier"]

[tool.ruff]
src = ["src", "tests"]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]

[tool.ruff.lint.isort]
known-first-party = ["fediaf_verifier"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "standard"
venvPath = "."
venv = ".venv"
include = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
markers = [
    "slow: marks tests that call the Anthropic API (deselect with '-m \"not slow\"')",
]
